requirements.txt
gymnasium
numpy
matplotlib
scipy
torch
tqdm
stable-baselines3